requirements.txt
flask
qrcode
pillow
gunicorn
urllib
